refacto de zinzin

See merge request pluff/beepenmieux!28

+++ src/message/message.controller.ts
@@ -1,20 +1,21 @@
 import {
-  Controller,
-  Get,
-  Post,
   Body,
-  Patch,
-  Param,
+  Controller,
   Delete,
-  Headers,
+  Get,
   HttpException,
+  Param,
+  Patch,
+  Post,
+  Req,
 } from '@nestjs/common';
-import { MessageService } from './message.service';
+import { AuthService } from 'src/auth/auth.service';
+import { GroupeService } from 'src/groupe/groupe.service';
+import { UsersService } from 'src/users/users.service';
 import { CreateMessageDto } from './dto/create-message.dto';
 import { UpdateMessageDto } from './dto/update-message.dto';
-import { AuthService } from 'src/auth/auth.service';
-import { UsersService } from 'src/users/users.service';
-import { GroupeService } from 'src/groupe/groupe.service';
+import { MessageService } from './message.service';
+import { User } from '@prisma/client';
 
 @Controller('message')
 export class MessageController {
@@ -30,37 +31,51 @@ export class MessageController {
     return this.messageService.create(createMessageDto);
   }
 
-  @Get()
-  findAll() {
-    return this.messageService.findAll();
-  }
-
   @Get(':id')
-  findOne(@Param('id') id: number) {
-    return this.messageService.findOne(id);
+  async findOne(@Param('id') id: number, @Req() request: Request) {
+    // check if user is in groupe
+    const user = request['user'];
+    const message = await this.messageService.findOne(id);
+    const groupe = await this.groupeService.findGroupesById(message.groupeId);
+
+    if ((await this.userService.isInGroupe(user, groupe)) === false) {
+      throw new HttpException('Unauthorized', 401);
+    }
+
+    if (!message) {
+      throw new HttpException('Message not found', 404);
+    }
+
+    return message;
   }
 
   @Get('groupe/:id')
-  async findAllByGroup(
-    @Param('id') id: string,
-    @Headers('authorization') authorization: string,
-  ) {
-    if (!authorization) {
-      throw new HttpException('Unauthorized', 401);
-    }
-    const token = authorization.split(' ')[1];
-    const userProfile = await this.authService.infoUser(token);
+  async findAllByGroup(@Param('id') id: string, @Req() request: Request) {
+    const userProfile = request['user'];
     const groupe = await this.groupeService.findByName(id);
-    if (
-      (await this.userService.isInGroupe(userProfile.user, groupe)) === false
-    ) {
+    if ((await this.userService.isInGroupe(userProfile, groupe)) === false) {
       throw new HttpException('Unauthorized', 401);
     }
     return this.messageService.findAllByGroup(id);
   }
 
   @Patch(':id')
-  update(@Param('id') id: number, @Body() updateMessageDto: UpdateMessageDto) {
+  async update(
+    @Param('id') id: number,
+    @Body() updateMessageDto: UpdateMessageDto,
+    @Req() request: Request,
+  ) {
+    const user: User = request['user'];
+    const message = await this.messageService.findOne(id);
+
+    if (!message) {
+      throw new HttpException('Message not found', 404);
+    }
+
+    if (user.id !== message.authorId) {
+      throw new HttpException('Unauthorized', 401);
+    }
+
     const messageUpdateInput = {
       ...updateMessageDto,
       author: {
@@ -74,7 +89,18 @@ export class MessageController {
   }
 
   @Delete(':id')
-  remove(@Param('id') id: number) {
+  async remove(@Param('id') id: number, @Req() request: Request) {
+    const user: User = request['user'];
+    const message = await this.messageService.findOne(id);
+
+    if (!message) {
+      throw new HttpException('Message not found', 404);
+    }
+
+    if (user.id !== message.authorId) {
+      throw new HttpException('Unauthorized', 401);
+    }
+
     return this.messageService.remove(id);
   }
 }
